backend/requirements.txt
# Backend Dependencies for StoryCircuit

# FastAPI Framework
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Azure SDKs
azure-ai-projects>=2.0.0b1
azure-identity>=1.15.0
azure-cosmos>=4.5.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0

# HTTP & Async
httpx>=0.26.0
aiofiles>=23.2.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
tenacity>=8.2.0

# Logging & Monitoring
structlog>=24.1.0
opencensus-ext-azure>=1.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Optional: Database migrations (if using SQL)
# alembic>=1.13.0
